looper check for just config samples
