+++ looper/utils.py
@@ -3,6 +3,7 @@
 from argparse import ArgumentParser
 from collections import defaultdict, Iterable
 import contextlib
+import glob
 import logging
 import os
 import random
@@ -104,6 +105,39 @@ def expandpath(path):
     :return str: expanded version of input path
     """
     return os.path.expandvars(os.path.expanduser(path)).replace("//", "/")
+
+
+
+def fetch_flag_files(prj):
+    """
+    Find all flag file paths for the given project.
+
+    :param Project | AttributeDict prj: full Project or AttributeDict with
+        similar metadata and access/usage pattern
+    :return Mapping[str, list[str]]: collection of filepaths associated with
+        particular flag for samples within the given project
+    """
+    # TODO: import from pep.
+    flags = ["completed", "running", "failed", "waiting", "partial"]
+
+    # Just create the filenames once, and pair once with flag name.
+    files = ["{}.flag".format(f) for f in flags]
+    flag_file_pairs = list(zip(flags, files))
+
+    # Collect the flag file paths by flag name.
+    files_by_flag = defaultdict(list)
+
+    # Iterate over samples to collect flag files.
+    for s in prj.samples:
+        folder = sample_folder(prj, s)
+
+        # Check each candidate flag for existence, collecting it if present.
+        for flag, flag_file in flag_file_pairs:
+            flag_path = os.path.join(folder, flag_file)
+            if os.path.isfile(flag_path):
+                files_by_flag[flag].append(flag_path)
+    
+    return files_by_flag
 
 
 
@@ -265,6 +299,18 @@ def partition(items, test):
         group = passes if test(item) else fails
         group.append(item)
     return passes, fails
+
+
+
+def sample_folder(prj, sample):
+    """
+    Get the path to this Project's root folder for the given Sample.
+
+    :param AttributeDict | Project prj: project with which sample is associated
+    :param Sample sample: the Sample for which to get root folder path
+    :return str: this Project's root folder for the given Sample
+    """
+    return os.path.join(prj.metadata.results_subdir, sample.name)
